requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.2
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cssselect2==0.4.1
defusedxml==0.7.1
demjson==2.2.4
diff-match-patch==20200713
Django==3.1.7
django-import-export==2.5.0
django-renderpdf==3.0.1
django-widget-tweaks==1.4.8
et-xmlfile==1.0.1
future==0.18.2
fuzzywuzzy==0.18.0
gunicorn==20.1.0
html5lib==1.1
idna==2.10
MarkupPy==1.14
mysqlclient @ file:///G:/Hind%20Bhoomi/hindbhoomi/mysqlclient-1.4.6-cp38-cp38-win32.whl
numpy==1.20.2
odfpy==1.4.1
olefile==0.46
openpyxl==3.0.7
pandas==1.2.3
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
pyinrail==1.0.0
PyPDF2==1.26.0
Pyphen==0.10.0
pytesseract==0.3.7
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
reportlab==3.5.67
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
tinycss2==1.1.0
urllib3==1.21.1
WeasyPrint==52.5
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
